tweaks to karabiner and nvim deployment

+++ nvim/setup_init.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
-# Neovim
 if [ ! -d $HOME/.config/nvim ]; then
     mkdir -p $HOME/.config/nvim;
 fi
 
-ln -s $HOME/git/dotfiles/nvim/init.lua $HOME/.config/nvim
+if ! [ -e "$HOME/.config/nvim/init.lua" ]; then
+    ln -s $HOME/git/dotfiles/nvim/init.lua $HOME/.config/nvim
+fi
